Enforce color in the test output
Turns out by default elm-test does not use ansi escape sequences when
invoked in a sub-shell.

That's clever. The test runner uses chalk.js to detect if the terminal
has colour support (roughly here:)
https://github.com/rtfeldman/node-test-runner/blob/master/lib/Generate.js#L3

and if so the application doesn't send the escape sequence.

In chalk documentation we can find that the color flag should be enough
to show the correct color to the terminal:
https://github.com/rtfeldman/node-test-runner/blob/master/lib/Generate.js#L3