Fix colouring for elm compilation errors
Instead of passing --color to elm-test I need to fake a tty
because elm-test runs the compiler from here:

https://github.com/rtfeldman/node-test-runner/blob/master/lib/Compile.js#L41

I believe the elm binary does not accept the color and due to that
colouring won't work on compilation errors.

By 'faking' a tty I can force the elm compiler and elm-test to
use colours as expected.